Revert "First steps into Idea support (#21)"
This reverts commit 83ca62586ebd35f9d126b40984e69e40ca762b0e.
