+++ .gitignore
@@ -132,6 +132,3 @@ reports/
 target/
 
 .bash_profile
-
-compiler-plugin/src/main/kotlin/arrow/meta/services/
-java_pid*
