+++ compiler-plugin/src/main/kotlin/arrow/meta/MetaCliProcessor.kt
@@ -6,7 +6,7 @@ import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
 //@AutoService(CommandLineProcessor::class)
 class MetaCliProcessor : CommandLineProcessor {
 
-  override val pluginId: String = "arrow.meta.compiler.plugin"
+  override val pluginId: String = "arrow.meta.plugin.compiler"
 
   override val pluginOptions: Collection<CliOption> = emptyList()
