+++ compiler-plugin/src/main/kotlin/arrow/meta/MetaPlugin.kt
@@ -1,7 +1,6 @@
 package arrow.meta
 
 import arrow.meta.comprehensions.comprehensions
-import arrow.meta.autofold.autoFold
 import arrow.meta.extensions.ExtensionPhase
 import arrow.meta.extensions.MetaComponentRegistrar
 import arrow.meta.higherkind.higherKindedTypes
@@ -11,7 +10,6 @@ import kotlin.contracts.ExperimentalContracts
 class MetaPlugin : MetaComponentRegistrar {
   override fun intercept(): List<ExtensionPhase> =
     higherKindedTypes +
-      autoFold +
       typeClasses +
       comprehensions
 }
