+++ compiler-plugin/src/main/kotlin/arrow/meta/extensions/MetaCommandLineProcessor.kt
@@ -5,7 +5,7 @@ import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
 
 class MetaCommandLineProcessor : CommandLineProcessor {
 
-  override val pluginId: String = "arrow.meta.compiler.plugin"
+  override val pluginId: String = "arrow.meta.plugin.compiler"
 
   override val pluginOptions: Collection<CliOption> = emptyList()
