+++ compiler-plugin/src/main/kotlin/arrow/meta/qq/ClassOrObject.kt
@@ -1,15 +1,11 @@
 package arrow.meta.qq
 
-import arrow.meta.utils.SealedSubclass
-import arrow.meta.utils.sealedSubclasses
-import org.jetbrains.kotlin.cli.common.messages.MessageCollector
+import arrow.meta.kt.renderSuperTypes
+import arrow.meta.kt.renderTypeParametersWithVariance
+import arrow.meta.kt.renderValueParameters
 import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.psi.KtClass
-import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.kotlin.psi.KtElement
-import org.jetbrains.kotlin.psi.KtNamedDeclaration
-import org.jetbrains.kotlin.psi.KtNamedFunction
-import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.psiUtil.getValueParameters
 import org.jetbrains.kotlin.psi.psiUtil.modalityModifierType
 import org.jetbrains.kotlin.psi.psiUtil.visibilityModifierType
@@ -24,12 +20,7 @@ interface ClassOrObject : Quote<KtElement, KtClass, ClassOrObject.ClassScope> {
     val typeParameters: Name,
     val valueParameters: Name,
     val supertypes: Name,
-    val sealedVariants: List<SealedSubclass>,
-    val body: Name,
-    val messageCollector: MessageCollector,
-    val functions: List<Name>,
-    val valueParameterNames: List<Name>,
-    val properties: List<Name>
+    val body: Name
   )
 
   override fun transform(ktElement: KtClass): ClassScope =
@@ -42,12 +33,7 @@ interface ClassOrObject : Quote<KtElement, KtClass, ClassOrObject.ClassScope> {
       valueParameters = if (ktElement.isInterface()) Name.identifier("")
       else Name.identifier(ktElement.renderValueParameters()),
       supertypes = Name.identifier(ktElement.renderSuperTypes()),
-      body = Name.identifier(ktElement.body?.text?.drop(1)?.dropLast(1).orEmpty()),
-      functions = ktElement.renderFunctions().map { it.nameAsSafeName },
-      properties = ktElement.renderProperties().map { it.nameAsSafeName },
-      valueParameterNames = ktElement.renderPrimaryParameters().map { it.nameAsSafeName },
-      sealedVariants = ktElement.sealedSubclasses(),
-      messageCollector = messageCollector()
+      body = Name.identifier(ktElement.body?.text?.drop(1)?.dropLast(1).orEmpty())
     )
 
   override fun KtClass.cleanUserQuote(quoteDeclaration: String): String =
@@ -58,29 +44,6 @@ interface ClassOrObject : Quote<KtElement, KtClass, ClassOrObject.ClassScope> {
 
   override fun parse(template: String): KtClass =
     quasiQuoteContext.compilerContext.ktPsiElementFactory.createClass(template)
-
-  fun KtClass.renderValueParameters(): String =
-    if (getValueParameters().isEmpty()) ""
-    else getValueParameters().joinToString(separator = ", ") { it.text }
-
-  fun messageCollector() =
-    quasiQuoteContext.compilerContext.messageCollector
-
-  fun KtClass.renderSuperTypes(): String =
-    superTypeListEntries.map { it.text }.joinToString(", ")
-
-  fun KtClass.renderTypeParametersWithVariance(): String =
-    typeParameters.joinToString(separator = ", ") { it.text }
-
-  fun KtClassOrObject.renderFunctions(): List<KtNamedDeclaration> =
-    declarations.filterIsInstance<KtNamedFunction>()
-
-  fun KtClassOrObject.renderProperties(): List<KtNamedDeclaration> =
-    declarations.filterIsInstance<KtProperty>()
-
-  fun KtClassOrObject.renderPrimaryParameters(): List<KtNamedDeclaration> =
-    primaryConstructorParameters.filter { it.hasValOrVar() }.fold(emptyList())
-    { acc, ktParameter -> acc + (ktParameter as KtNamedDeclaration) }
 
   companion object : Quote.Factory<KtElement, KtClass, ClassScope> {
     override operator fun invoke(
@@ -96,4 +59,5 @@ interface ClassOrObject : Quote<KtElement, KtClass, ClassOrObject.ClassScope> {
         override val containingDeclaration: KtElement = containingDeclaration
       }
   }
+
 }
