+++ core/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'kotlin'
-apply plugin: 'maven-publish'
 
 def compilerPlugin = project(':compiler-plugin')
