+++ core/src/main/kotlin/arrow/core/TestClass.kt
@@ -1,9 +1,9 @@
 package arrow.core
 
 //import arrow.core.ForOption //works because in the same module
+
+//interface Kind<out F, out A>
 //
-////interface Kind<out F, out A>
-////
-////sealed class Option<out A>
-////object None : Option<Nothing>()
-////data class Some<out A>(val a: A) : Option<A>()
+//sealed class Option<out A>
+//object None : Option<Nothing>()
+//data class Some<out A>(val a: A) : Option<A>()
