+++ gradle-plugin/build.gradle
@@ -10,20 +10,22 @@ project.group = 'io.arrow-kt'
 project.version = '0.0.1'
 
 dependencies {
-  compile gradleApi()
-  compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-  compile "org.jetbrains.kotlin:kotlin-gradle-plugin-api:$kotlin_version"
-  compile "org.jetbrains.kotlin:kotlin-compiler-embeddable:$kotlin_version"
+  implementation gradleApi()
+  implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+  implementation "org.jetbrains.kotlin:kotlin-gradle-plugin-api:$kotlin_version"
   compileOnly project(":compiler-plugin") { "embeddable" }
 }
 
 compileKotlin {
   kotlinOptions.jvmTarget = "1.8"
 }
+compileTestKotlin {
+  kotlinOptions.jvmTarget = "1.8"
+}
 
 jar {
   // Embed compiler plugin in jar
-  from(zipTree(configurations.named("compileClasspath").get().find { it.name.startsWith("compiler-plugin") }))
+  configurations.named("compileClasspath").get().find { it.name.startsWith("plugin-compiler") }
 
   manifest {
     attributes["Specification-Title"] = project.name
@@ -36,7 +38,7 @@ jar {
 gradlePlugin {
   plugins {
     arrow {
-      id = "arrow.meta.compiler.plugin"
+      id = "arrow.meta.plugin.gradle"
       implementationClass = "arrow.meta.plugin.gradle.ArrowGradlePlugin"
     }
   }
@@ -50,12 +52,12 @@ pluginBundle {
 
   plugins {
     arrow {
-      id = "arrow.meta.compiler.plugin"
+      id = "arrow.meta.plugin.gradle"
       displayName = "Arrow compiler plugin"
     }
   }
 
   mavenCoordinates {
-    groupId = "arrow.meta.compiler.plugin"
+    groupId = "arrow.meta.plugin.gradle"
   }
 }
