+++ idea-plugin/src/main/kotlin/arrow/meta/plugin/idea/ArrowImportHandler.kt
@@ -37,14 +37,13 @@ internal object ArrowImportHandler {
 
     // Remove the incompatible compiler plugin from the classpath if found
     var isEnabled = false
-    val oldPluginClasspaths = (commonArguments.pluginClasspaths ?: emptyArray()).filterTo(mutableListOf()) {
-      val match = regex.matches(it) && validateJar(it)
-      logger.debug("$it [$match]")
-      if (match) {
-        isEnabled = true
+    val oldPluginClasspaths =
+      (commonArguments.pluginClasspaths ?: emptyArray()).filterTo(mutableListOf()) {
+        val match = regex.matches(it) && validateJar(it)
+        logger.debug("$it [$match]")
+        isEnabled = match
+        !match
       }
-      !match
-    }
 
     // Add the compatible compiler plugin version to the classpath if available and is enabled in Gradle
     val newPluginClasspaths = if (isEnabled && PLUGIN_JPS_JAR != null)
@@ -67,7 +66,7 @@ internal object ArrowImportHandler {
       val jar = JarInputStream(FileInputStream(path))
       val manifest = jar.manifest
       manifest.mainAttributes.getValue("Implementation-Title")
-        .startsWith("arrow.meta.compiler.plugin")
+        .startsWith("arrow.meta.plugin.gradle")
     } catch (_: Exception) {
       false
     }
