+++ idea-plugin/src/main/resources/META-INF/plugin.xml
@@ -1,10 +1,13 @@
 <idea-plugin>
-<id>arrow.meta.compiler.plugin</id>
+<id>arrow-plugin</id>
 <name>Arrow</name>
 
 <vendor email="support@yourcompany.com" url="http://www.yourcompany.com">YourCompany</vendor>
 
-<description>The Arrow Intellij IDEA plugin.</description>
+<description><![CDATA[
+    The Arrow Intellij IDEA plugin.<br>
+    <em>most HTML tags may be used</em>
+    ]]></description>
 
 <!-- please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/plugin_compatibility.html
      on how to target different products -->
